+++ lib/components/MarkedText/hooks/useFocus/hooks/useKeyDown.tsx
@@ -1,10 +1,12 @@
 import {KeyboardEvent, MutableRefObject, RefObject, useEffect} from "react";
 import {Caret} from "../../../../../utils/Caret";
 import {KEY} from "../../../../../constants";
-import {KeyedPieces, Type} from "../../../../../types";
-import {genHash, useStore} from "../../../../../utils";
+import {Type} from "../../../../../types";
+import {useStore} from "../../../../../utils";
 import {Recovery} from "./useRecoveryAfterRemove";
+import {Oracle} from "../../../../../utils/Oracle";
 
+//TODO clean up
 export function useKeyDown(
     spanRefs: Map<number, RefObject<HTMLElement>>,
     recoveryRef: MutableRefObject<Recovery | null>,
@@ -61,113 +63,4 @@ export function useKeyDown(
             }
         }
     ), [pieces])
-}
-
-class Oracle {
-    constructor(
-        readonly focusedSpanRef: MutableRefObject<HTMLElement | null>,
-        readonly spanRefs: Map<number, RefObject<HTMLSpanElement>>,
-        readonly recoveryRef: MutableRefObject<Recovery | null>,
-        readonly pieces: KeyedPieces
-    ) {
-    }
-
-    recoverLeft() {
-        let previousText = this.getPreviousElement()?.textContent!
-        let currentText = this.focusedSpanRef.current?.textContent!
-        let newText = previousText + currentText
-
-        this.recoveryRef.current = {
-            key: genHash(newText),
-            caret: previousText.length
-        }
-    }
-
-    recoverRight() {
-        let nextText = this.getNextElement()?.textContent!
-        let currentText = this.focusedSpanRef.current?.textContent!
-        let newText = currentText + nextText
-
-        this.recoveryRef.current = {
-            key: genHash(newText),
-            caret: currentText.length
-        }
-    }
-
-    getPreviousElement() {
-        let previous: HTMLElement | null = null
-
-        if (!this.focusedSpanRef.current)
-            return null
-
-        for (const [key, elementRef] of this.spanRefs) {
-            if (elementRef.current === this.focusedSpanRef.current)
-                return previous
-
-            previous = elementRef.current
-        }
-
-        return null
-    }
-
-    getNextElement() {
-        let isNext = false
-
-        if (!this.focusedSpanRef.current)
-            return null
-
-        for (const [key, elementRef] of this.spanRefs) {
-            if (isNext)
-                return elementRef.current
-
-            if (elementRef.current === this.focusedSpanRef.current)
-                isNext = true
-        }
-
-        return null
-    }
-
-    getPreviousMarkKey() {
-        let previous: number | null = null
-        const current = this.getCurrentKey()
-
-        if (!current) return null
-
-        for (const [key] of this.pieces) {
-            if (key === current)
-                return previous
-
-            previous = key
-        }
-
-        return null
-    }
-
-    getNextMarkKey() {
-        let isNext = false
-        const current = this.getCurrentKey()
-
-        if (!current) return null
-
-        for (const [key] of this.pieces) {
-            if (isNext)
-                return key
-
-            if (key === current)
-                isNext = true
-        }
-
-        return null
-    }
-
-    getCurrentKey() {
-        if (!this.focusedSpanRef.current) return null
-
-        for (const [key, elementRef] of this.spanRefs) {
-            if (elementRef.current === this.focusedSpanRef.current)
-                return key
-        }
-
-        return null
-    }
 }
